rename to signer, remove chan receipt typedef
