+++ relay/nakama/allowlist.go
@@ -9,7 +9,7 @@ import (
 	"fmt"
 	"os"
 	nakamaerrors "pkg.world.dev/world-engine/relay/nakama/errors"
-	"pkg.world.dev/world-engine/relay/nakama/pk"
+	"pkg.world.dev/world-engine/relay/nakama/signer"
 	"pkg.world.dev/world-engine/relay/nakama/utils"
 	"strconv"
 	"strings"
@@ -72,7 +72,7 @@ func allowListRPC(ctx context.Context, logger runtime.Logger, _ *sql.DB, nk runt
 	if err != nil {
 		return utils.LogErrorFailedPrecondition(logger, err)
 	}
-	if id != pk.AdminAccountID {
+	if id != signer.AdminAccountID {
 		return utils.LogError(
 			logger,
 			eris.Errorf("unauthorized: only admin may call this RPC"),
@@ -205,7 +205,7 @@ func writeVerified(ctx context.Context, nk runtime.NakamaModule, userID string) 
 		{
 			Collection:      allowedUsers,
 			Key:             userID,
-			UserID:          pk.AdminAccountID,
+			UserID:          signer.AdminAccountID,
 			Value:           string(bz),
 			Version:         "",
 			PermissionRead:  runtime.STORAGE_PERMISSION_NO_READ,
@@ -225,7 +225,7 @@ func isUserVerified(ctx context.Context, nk runtime.NakamaModule, userID string)
 		{
 			Collection: allowedUsers,
 			Key:        userID,
-			UserID:     pk.AdminAccountID,
+			UserID:     signer.AdminAccountID,
 		},
 	})
 	if err != nil {
@@ -242,7 +242,7 @@ func readKey(ctx context.Context, nk runtime.NakamaModule, key string) (*KeyStor
 		{
 			Collection: allowlistKeyCollection,
 			Key:        key,
-			UserID:     pk.AdminAccountID,
+			UserID:     signer.AdminAccountID,
 		},
 	})
 	if err != nil {
@@ -270,7 +270,7 @@ func writeKey(ctx context.Context, nk runtime.NakamaModule, ks *KeyStorage) erro
 		{
 			Collection:      allowlistKeyCollection,
 			Key:             ks.Key,
-			UserID:          pk.AdminAccountID,
+			UserID:          signer.AdminAccountID,
 			Value:           string(bz),
 			Version:         "",
 			PermissionRead:  runtime.STORAGE_PERMISSION_NO_READ,
