+++ relay/nakama/signer/private_key_test.go
@@ -1,4 +1,4 @@
-package pk
+package signer
 
 import (
 	"context"
